[autoparallel] update CommSpec
